Reading activities from 7z packages is quite fast
Parse package ...
  ... parsed 1000 data sets
  ... parsed 2000 data sets
  ... parsed 3000 data sets
  ... parsed 4000 data sets
  ... parsed 5000 data sets
  ... parsed 6000 data sets
  ... parsed 7000 data sets
  ... parsed 8000 data sets
  ... parsed 9000 data sets
  ... parsed 10000 data sets
  ... parsed 11000 data sets
  ... parsed 12000 data sets
  ... parsed 13000 data sets
  ... parsed 14000 data sets
  ... parsed 15000 data sets
  ... parsed 16000 data sets
  ... parsed 17000 data sets
  ... parsed 18000 data sets
  ... parsed 19000 data sets
  ... parsed 20000 data sets
  ... parsed 21000 data sets
  ... parsed 22000 data sets
  ... parsed 23000 data sets
activities              | 23523
intermediate exchanges  | 4057
elementary exchanges    | 4362
units                   |   130
unit conversions        |   774
persons                 |   456
sources                 |  1395
parsed package in 58.41 seconds

+++ src/main/java/spold2/Spold2.java
@@ -65,10 +65,10 @@ public class Spold2 {
 		IntermediateExchange candidate = null;
 		for (IntermediateExchange e : getProducts(ds)) {
 			Integer og = e.outputGroup;
-			if (og == null || og.intValue() != 0)
+			if (og == null || og != 0)
 				continue;
 			Double a = e.amount;
-			if (a != null && a.doubleValue() != 0)
+			if (a != null && a != 0)
 				return e;
 			candidate = e;
 		}
